Rewrote views to handle all CRUD operations using one view.

+++ apps/calendar/views.py
@@ -11,42 +11,29 @@ def calendar(request):
 
 
 @json
-def calendar_read(request):
+def calendars(request):
   assert request.method == "GET"
-  if 'id' in request.GET:
-    calendar = Calendar.objects.get(id=int(request.GET['id']))
-    return calendar.to_json()
-  else:
-    calendars = Calendar.objects.all()
-    return [calendar.to_json() for calendar in calendars]
-
-
-def task_create(request):
-  assert request.method == 'POST'
-  Task.objects.create(**request.POST)
-  return HttpResponse()
-
+  if request.method == "GET":
+    if 'id' in request.GET:
+      calendar = Calendar.objects.get(id=int(request.GET['id']))
+      return calendar.to_json()
+    else:
+      calendars = Calendar.objects.all()
+      return [calendar.to_json() for calendar in calendars]
 
 @json
-def task_read(request):
-  assert request.method == 'GET'
-  if 'id' in request.GET:
-    task = Task.objects.get(id=int(request.GET['id']))
-    return task.to_json()
-  else:
-    tasks = Task.objects.all()
-    return [task.to_json() for task in tasks]
-    
-
-
-def task_update(request):
-  assert request.method == 'POST'
-  Task.objects.update(**request.POST)
-  return HttpResponse()
-
-
-def task_destroy(request):
-  assert request.method == 'POST'
-  assert 'id' in request.POST
-  Task.objects.get(id=int(request.POST['id'])).delete()
-  return HttpResponse()
+def tasks(request):
+  if request.method == "POST":
+    return Task.objects.create(**request.POST)
+  elif request.method == "GET":
+    if 'id' in request.GET:
+      task = Task.objects.get(id=int(request.GET['id']))
+      return task.to_json()
+    else:
+      tasks = Task.objects.all()
+      return [task.to_json() for task in tasks]
+  elif request.method == "PUT":
+    return Task.objects.update(**request.POST)
+  elif request.method == "DESTROY":
+    Task.objects.get(id=int(request.POST['id'])).delete()
+    return True
